Change base image for docker image to ubuntu 16.04
Image based on debian 9 or alpine have some kind of bug that make gormun
crash in libzmq. It might be caused by the version of libzmq, or another
dependency. I did not easily reproduce this issue locally with docker, it
only appear in production...

For now we will use a configuration known to work while we analyse what
happens in these environments